+++ src/libkefir_proggen.h
@@ -13,6 +13,6 @@ struct kefir_cprog *
 proggen_make_cprog_from_filter(const struct kefir_filter *filter,
 			       const struct kefir_cprog_attr *attr);
 int proggen_cprog_to_buf(const struct kefir_cprog *cprog, char **buf,
-			 size_t *buf_len);
+			 unsigned int *buf_len);
 
 #endif /* LIBKEFIR_PROGGEN_H */
